Move icc profile method to metadata
